% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.tf}
\alias{autoplot.tf}
\alias{autolayer.tf}
\title{Autoplot and autolayer methods for \code{tf} objects}
\usage{
\method{autoplot}{tf}(object, ...)

\method{autolayer}{tf}(object, ...)
}
\arguments{
\item{object}{a \code{tf} object}

\item{...}{passed to \code{\link[ggplot2:geom_line]{geom_line()}}}
}
\value{
A \code{\link[=tf_ggplot]{tf_ggplot()}} object for \code{autoplot()}, a \code{\link[ggplot2:layer]{ggplot2::layer()}} object for \code{autolayer()}.
}
\description{
Convenient plotting methods for \code{tf} objects. \code{autoplot()} creates a
complete spaghetti plot, \code{autolayer()} creates a layer that can be
added to an existing \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} or \code{\link[=tf_ggplot]{tf_ggplot()}}.
}
\examples{
\donttest{
library(ggplot2)
f <- tf_rgp(5)
autoplot(f)
ggplot() + autolayer(f)
tf_ggplot() + autolayer(f)
}
}
\seealso{
Other tidyfun visualization:
\code{\link[=autoplot.tf_mv]{autoplot.tf_mv()}},
\code{\link{ggcapellini}},
\code{\link[=gglasagna]{gglasagna()}},
\code{\link{ggspaghetti}},
\code{\link[=tf_phaseplane]{tf_phaseplane()}}
}
\concept{tidyfun visualization}
