% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{convert_tf_ggplot}
\alias{convert_tf_ggplot}
\title{Convert tf_ggplot object to regular ggplot with transformed data}
\usage{
convert_tf_ggplot(tf_plot, layer_mapping = aes())
}
\arguments{
\item{tf_plot}{A tf_ggplot object}

\item{layer_mapping}{Additional aesthetic mapping from a layer}
}
\value{
Regular ggplot object with transformed data
}
\description{
Convert tf_ggplot object to regular ggplot with transformed data
}
\keyword{internal}
