% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth}
\alias{depth}
\alias{depth.matrix}
\alias{depth.fvector}
\title{Data Depth}
\usage{
depth(x, depth = "MBD", ...)

\method{depth}{matrix}(x, depth = "MBD", argvals = seq_len(ncol(x)), ...)

\method{depth}{fvector}(x, depth = "MBD", argvals = tidyfun::argvals(x),
  ...)
}
\arguments{
\item{x}{\code{fvector} (or a matrix of evaluations)}

\item{depth}{currently available: "MBD", i.e. modified band depth}

\item{argvals}{grid of evaluation points}
}
\value{
vector of depth values
}
\description{
Data depth for functional data. Currently implemented: Modified Band Depth
}
\references{
Sun, Y., Genton, M. G., & Nychka, D. W. (2012).
Exact fast computation of band depth for large functional datasets:
How quickly can one million curves be ranked?. \emph{Stat}, \strong{1}(1), 68-74.
Lopez-Pintado, S. and Romo, J. (2009).
On the Concept of Depth for Functional Data.
\emph{Journal of the American Statistical Association}, \strong{104}, 718-734.
}
