% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.tfd}
\alias{evaluate.tfb}
\alias{evaluate.data.frame}
\title{Evaluate \code{tf}s, both inside or outside a \code{data.frame}}
\usage{
evaluate(object, arg, ...)

\method{evaluate}{tfd}(object, arg, ...)

\method{evaluate}{tfb}(object, arg, ...)

\method{evaluate}{data.frame}(object, arg, ...)
}
\arguments{
\item{object}{an \code{tf} or a \code{data.frame}-like object with \code{tf} columns}

\item{arg}{optional evaluation grid, defaults to \code{arg(object)}.}

\item{...}{optional: names of the \code{tf}-columns to unnest}
}
\description{
The \code{evaluate.data.frame} method evaluates \code{tf}-columns inside a \code{data.frame}
into list columns of smaller \code{data.frames} containing the functions' arguments
(\code{arg}) and evaluations (\code{value}). Its \code{arg}-argument can be a list of \code{arg}-vectors
used as the \code{arg} argument for the \code{\link[=evaluate]{evaluate()}}-method for the respective
\code{tf}-columns in \code{object}.
}
\seealso{
\code{?`[.tf`}
}
