% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feval-def.R
\name{feval}
\alias{feval}
\alias{feval.matrix}
\alias{feval.data.frame}
\alias{feval.list}
\title{Constructors for functional data evaluated on grids of argument values}
\usage{
feval(data, ...)

\method{feval}{matrix}(data, argvals = NULL, regular = NULL,
  domain = NULL, range = NULL, evaluator = approx_linear, signif = 4,
  ...)

\method{feval}{data.frame}(data, id = 1, argvals = 2, value = 3,
  domain = NULL, range = NULL, evaluator = approx_linear, signif = 4,
  ...)

\method{feval}{list}(data, argvals = NULL, regular = NULL, domain = NULL,
  range = NULL, evaluator = approx_linear, signif = 4, ...)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} or \code{list} of suitable shape.}

\item{...}{not used}

\item{argvals}{}
}
\value{
an \code{feval}-object
}
\description{
Various constructor methods
}
