% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{finalize_tf_ggplot}
\alias{finalize_tf_ggplot}
\title{Finalize tf_ggplot by processing all tf expressions and creating regular ggplot}
\usage{
finalize_tf_ggplot(tf_plot)
}
\arguments{
\item{tf_plot}{A tf_ggplot object with accumulated layers}
}
\value{
Regular ggplot object with all layers properly transformed
}
\description{
Finalize tf_ggplot by processing all tf expressions and creating regular ggplot
}
\keyword{internal}
