% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb-fpc-utils.R
\name{fpc_wsvd}
\alias{fpc_wsvd}
\title{Eigenfunctions via weighted SVD}
\usage{
fpc_wsvd(data, pve = 0.995)
}
\arguments{
\item{data}{mueric matrix of function evaluations (each row is one curve, no NAs)}

\item{pve}{percentage of variance explained for truncation lag}

\item{arg}{numeric vector of argument values}
}
\value{
a list with entries
\itemize{
\item `mu`` estimated mean function (numeric vector)
\item `efunctions`` estimated FPCs (numeric matrix, columns represent FPCs)
\item \code{scores} estimated FPC scores (one row per observed curve)
\item \code{npc} how many FPCs were returned for the given \code{pve} (integer)
}
}
\description{
Compute (truncated) orthonormal eigenfunctions and scores
for data potentially on a non-equidistant grid.
}
\references{
code adapted from / inspired by \code{wsvd()} function of Bioconductor
package \code{mogsa} by Cheng Meng.
}
