% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbase-fpc.R
\name{fpcbase}
\alias{fpcbase}
\alias{fpcbase.data.frame}
\alias{fpcbase.matrix}
\alias{fpcbase.numeric}
\alias{fpcbase.fvector}
\title{Constructors/converters for functional data in FPC-basis representation}
\usage{
fpcbase(data, ...)

\method{fpcbase}{data.frame}(data, id = 1, argvals = 2, value = 3,
  domain = NULL, smooth = TRUE, signif = 4, ...)

\method{fpcbase}{matrix}(data, argvals = NULL, domain = NULL,
  smooth = TRUE, signif = 4, ...)

\method{fpcbase}{numeric}(data, argvals = NULL, domain = NULL,
  smooth = TRUE, signif = 4, ...)

\method{fpcbase}{fvector}(data, argvals = NULL, smooth = TRUE, ...)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} or \code{list} of suitable shape, or another \code{fvector}-object.}

\item{...}{arguments to the call to \code{\link[refund:fpca.sc]{refund::fpca.sc()}} that computes the
(regularized/smoothed) FPCA. Unless set by the user \code{tidyfun} uses \code{pve = .995} to
determine the truncation levels and uses \code{bs = 15} basis functions for
the mean function and the marginal bases for the covariance surface.}

\item{id}{The name/number of the column defining which data belong to which function.}

\item{argvals}{\code{numeric}, or list of \code{numeric}s. The evaluation grid. See Details.
For the \code{data.frame}-methods: the name/number of the column defining the evaluation grid.}

\item{value}{The name/number of the column containing the function evaluations.}

\item{domain}{range of the \code{argvals}.}

\item{smooth}{use smoothed mean function and smoothed covariance surface estimates
from \code{\link[refund:fpca.sc]{refund::fpca.sc()}} or simply perform a (truncated) PCA of the data? See Details.}

\item{signif}{significant digits of the "resolution" of the evaluation grid.  See details.}
}
\description{
These functions perform a (functional) principal component analysis of the
input data and return an \code{fbase} \code{fvector}-object that uses the empirical
eigenfunctions as basis functions for representing the data. By default, a
\code{smooth}ed FPCA via \code{\link[refund:fpca.sc]{refund::fpca.sc()}} is used to compute eigenfunctions and
scores based on the smoothed empirical covariance.
If \code{smooth =FALSE}, a fast, unregularized PCA of the data is done instead.
}
\details{
ATM, the unsmoothed version does not use orthogonal basis "functions"
for irregular / non-equidistant grids and does not work for
incomplete/irregular data.
}
\seealso{
fbase
}
