% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funplot.R
\name{funplot}
\alias{funplot}
\title{Quick \code{ggplot2}-plots for \code{tf}-data}
\usage{
funplot(
  f,
  arg,
  n_grid = 50,
  points = is_irreg(f),
  type = c("spaghetti", "lasagna"),
  alpha = min(1, max(0.05, 2/length(f)))
)
}
\arguments{
\item{f}{an \code{tf} object}

\item{arg}{(optional) numeric vector to be used as \code{arg}. See details.}

\item{n_grid}{minimal size of equidistant grid used for plotting,
defaults to 50. See details.}

\item{points}{should the original evaluation points be marked by points?
Defaults to \code{TRUE} for irregular \code{tfd} and FALSE for all others}

\item{type}{"spaghetti": line plots, "lasagna": heat maps.}

\item{alpha}{alpha-value (see\code{\link[grDevices:rgb]{grDevices::rgb()}}) for noodle transparency.
Defaults to 2/(no. of observations). Lower is more transparent.}
}
\value{
for \code{funplot}: the \code{ggplot}-object for further modification. For the
others: the plotted \code{tf}-object, invisibly.
}
\description{
\code{ggplot}-functions for displaying functional data in
spaghetti- (i.e., line plots) and lasagna- (i.e., heat map) flavors.
}
\details{
If no \code{arg} are provided, evaluation points (\code{arg}) for the functions
are given by the union of the \code{tf}'s \code{arg} and an equidistant grid
over its domain with \code{n_grid} points. If you want to only see the original
data for \code{tfd}-objects without inter-/extrapolation, use \code{n_grid < 1} or
\code{n_grid = NA}.
}
\seealso{
\code{\link[tf:tfviz]{tf::plot.tf()}}
}
