% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{deriv.feval}
\alias{deriv.feval}
\alias{deriv.fbase}
\alias{integrate}
\title{Derivatives and integrals of functional data}
\usage{
\method{deriv}{feval}(expr, order = 1, argvals = NULL, ...)

\method{deriv}{fbase}(expr, order = 1, ...)

integrate(f, lower, upper, ...)
}
\arguments{
\item{expr}{an \code{fvector}}

\item{order}{order of differentiation. Maximally 2 for \code{fbase} with \code{mgcv}-spline bases.}

\item{argvals}{grid to use for the differentiation. Not the \code{argvals} of the returned object.}

\item{...}{not used}
}
\value{
an \code{fvector} with (slightly) different \code{argvals} (and \code{basis})
}
\description{
Derivatives of \code{fvectors} use finite differences of the evaluations
for \code{feval} and finite differences of the basis functions for \code{fbase}.
For many types of bases, the latter seems to become rather unreliable
at the outer limits of the functions' domain, for some reason...
}
