% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvector-methods.R
\name{c.fvector}
\alias{c.fvector}
\alias{c.feval}
\alias{c.fbase}
\title{Concatenate \code{fvector}-objects}
\usage{
\method{c}{fvector}(...)

\method{c}{feval}(...)

\method{c}{fbase}(...)
}
\arguments{
\item{...}{a bunch of \code{fvector}-objects on the same domain and of the same class}
}
\value{
a new and longer \code{fvector}-object of the same class and with the same attributes as the
the first argument
}
\description{
Functions to concatenate multiple vectors of functional data.
}
\details{
Currently only allows concatenation of very similar functions (same class, same domain).
\code{c.fbase} will use the basis of its first argument for representing all remaining arguments.
\code{c.feval} will use the \code{evaluator} of its first argument for all remaining arguments as well.
}
