% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/utils.R
\name{argvals}
\alias{argvals}
\alias{evaluations}
\alias{n_evaluations}
\alias{domain}
\alias{evaluator}
\alias{basis}
\alias{evaluator<-}
\alias{argvals<-}
\alias{coef.fbase}
\alias{rev.fvector}
\alias{is_fvector}
\alias{is_irreg}
\alias{is_feval}
\alias{is_fbase}
\alias{is.na.fvector}
\title{Utility functions for \code{fvector}-objects}
\usage{
argvals(f)

evaluations(f)

n_evaluations(f)

domain(f)

evaluator(f)

basis(f)

evaluator(x) <- value

argvals(x) <- value

\method{coef}{fbase}(object, ...)

\method{rev}{fvector}(x)

is_fvector(x)

is_irreg(x)

is_feval(x)

is_fbase(x)

\method{is.na}{fvector}(x)
}
\arguments{
\item{f}{an \code{fvector} object}

\item{x}{an \code{fvector} object}

\item{value}{for \code{evaluator<-}: name of a function that can be used to interpolate an \code{feval}. Needs to
accept vector arguments \code{x}, \code{argvals}, \code{evaluations} and return
evaluations of the function defined by \code{argvals}, \code{evaluations} at \code{x}
for \code{argvals<-}: a list of grid points, for internal use only.}

\item{object}{as usual}

\item{...}{dots}
}
\description{
A bunch of methods & utilities that do what they say.
}
