% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-capellini.R
\name{ggcapellini}
\alias{ggcapellini}
\alias{StatCapellini}
\alias{stat_capellini}
\alias{geom_capellini}
\alias{geom_cappellini}
\alias{geom_capelini}
\alias{geom_cappelini}
\alias{GeomCapellini}
\title{Glyph plots for \code{tf} objects}
\usage{
stat_capellini(
  mapping = NULL,
  data = NULL,
  geom = "capellini",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  arg = NULL,
  add_lines = FALSE,
  add_boxes = TRUE,
  width = NULL,
  height = NULL,
  ...
)

geom_capellini(
  mapping = NULL,
  data = NULL,
  stat = "capellini",
  position = "identity",
  ...,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  arg = NULL,
  add_lines = TRUE,
  add_boxes = TRUE,
  width = NULL,
  height = NULL,
  box.colour = "#0000001A",
  box.linetype = 1,
  box.fill = NA,
  box.linewidth = 0.1,
  box.alpha = 0.1,
  line.colour = "black",
  line.linetype = 2,
  line.linewidth = 0.3,
  line.alpha = 0.5
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{remove NAs? defaults to \code{TRUE}}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{arg}{where to evaluate \code{tf}; defaults to the object's default
evaluation grid.}

\item{add_lines}{should a reference line in the middle of the range of the
functions' values be added to each glyph? defaults to TRUE}

\item{add_boxes}{should a box be added to frame each glyph? defaults to TRUE}

\item{width}{the width of the glyphs. Defaults to 2/3 of the \code{\link[ggplot2:resolution]{ggplot2::resolution()}}
of the variable for the \code{x}-aesthetic, this will be too small if any values
are close together.}

\item{height}{the height of the glyphs. Defaults to 2/3 of the \code{\link[ggplot2:resolution]{ggplot2::resolution()}}
of the variable for the \code{y}-aesthetic, this will be too small if any values
are close together.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{stat}{the stat to use; defaults to \code{"capellini"}.}

\item{box.colour}{aesthetic property of the box}

\item{box.linetype}{aesthetic property of the box}

\item{box.fill}{aesthetic property of the box}

\item{box.linewidth}{aesthetic property of the box}

\item{box.alpha}{aesthetic property of the box}

\item{line.colour}{aesthetic property of the reference line}

\item{line.linetype}{aesthetic property of the reference line}

\item{line.linewidth}{aesthetic property of the reference line}

\item{line.alpha}{aesthetic property of the reference line}
}
\value{
A \code{\link[ggplot2:layer]{ggplot2::layer()}} object for use in a ggplot.
}
\description{
Plots a miniature glyph / sparkline for each entry of a \code{tf}-object.
(Capellini are tiny spaghetti -- \emph{angel hair} pasta.) Aesthetics \code{x} and \code{y}
specify the location of the glyphs, the \code{tf} aesthetic defines their shapes.
The aliases \code{geom_cappelini}, \code{geom_cappellini}, and \code{geom_capelini} are
also accepted.
}
\section{\code{tf} aesthetic}{

Mandatory. Used to designate a column of class \code{tf} to be visualized as glyphs.
}

\examples{
\dontshow{if (rlang::is_installed(c("fda", "maps"))) withAutoprint(\{ # examplesIf}
library(ggplot2)
weather <- fda::CanadianWeather
canada <- data.frame(
  place = weather$place,
  region = weather$region,
  lat = weather$coordinates[, 1],
  lon = -weather$coordinates[, 2],
  region = weather$region
)
canada$temp <- tfd(t(weather$dailyAv[, , 1]), arg = 1:365)
canada$precipl10 <- tfd(t(weather$dailyAv[, , 3]), arg = 1:365) |> tf_smooth()
canada_map <-
  data.frame(maps::map("world", "Canada", plot = FALSE)[c("x", "y")])
# map of canada with annual temperature averages in red, precipitation in blue:
ggplot(canada, aes(x = lon, y = lat)) +
  geom_capellini(aes(tf = precipl10), width = 3, height = 5, colour = "blue") +
  geom_capellini(aes(tf = temp), width = 3, height = 5, colour = "red") +
  geom_path(data = canada_map, aes(x = x, y = y), alpha = 0.1) +
  coord_quickmap()

ggplot(canada, aes(x = lon, y = lat, colour = region)) +
  geom_capellini(aes(tf = precipl10),
    width = 5, height = 3,
    line.linetype = 1, box.fill = "white", box.alpha = 0.5, box.colour = NA
  )
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[GGally:glyphs]{GGally::glyphs()}}

Other tidyfun visualization:
\code{\link[=autoplot.tf]{autoplot.tf()}},
\code{\link[=autoplot.tf_mv]{autoplot.tf_mv()}},
\code{\link[=gglasagna]{gglasagna()}},
\code{\link{ggspaghetti}}
}
\concept{tidyfun visualization}
