% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\docType{data}
\name{ggtf}
\alias{ggtf}
\alias{StatTf}
\alias{stat_tf}
\alias{geom_spaghetti}
\alias{geom_meatballs}
\alias{geom_lasagna}
\title{Geoms for \code{tf} objects}
\usage{
stat_tf(mapping = NULL, data = NULL, geom = "line",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, arg = NULL, ...)

geom_spaghetti(mapping = NULL, data = NULL, position = "identity",
  na.rm = TRUE, show.legend = NA, inherit.aes = TRUE, arg = NULL, ...)

geom_meatballs(mapping = NULL, data = NULL, position = "identity",
  na.rm = TRUE, show.legend = NA, inherit.aes = TRUE, arg = NULL,
  spaghetti = TRUE, ...)

geom_lasagna(mapping = list(), data = NULL, position = "identity",
  na.rm = TRUE, show.legend = NA, inherit.aes = TRUE, arg = NULL,
  order_by = NULL, size = 4, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{remove NAs? defaults to \code{TRUE}}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{arg}{where to evaluate \code{tf} -- defaults to the default ;)}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{spaghetti}{plot noodles along with meatballs? defaults to true.}

\item{order_by}{a function that returns a single value when applied to a numeric vector,
used to define the vertical ordering of the functions in the plot (see Examples)}

\item{size}{width for lasagna layers. You will need to set this manually, sorry.}
}
\description{
Defines spaghetti and lasagna plot geoms for \code{tf}-objects.
\code{geom_spaghetti} does spaghetti plots, \code{geom_meatballs} does spaghetti plots
with points for the actual evaluations, \code{geom_lasagna} does lasagna plots.
All of this is a hack using \code{stat_tf} and subject to change.
}
\section{\code{tf} aesthetic}{

Mandatory. Used to designate a column of class \code{tf} to be visualized.
}

\section{\code{order} aesthetic}{

optional. Used to designate a variable that defines the (low-to-high) ordering of the
functions in the lasagna plot. See also: \code{order_by}
}

\examples{
set.seed(1221)
data = data.frame(col = sample(gl(5, 2)))
data$f = rgp(10)
data$fi = jiggle(data$f)
data$fb = tfb(data$f)
library(ggplot2)
ggplot(data, aes(tf = f, color = depth(f))) + geom_spaghetti() + 
  annotate("text", x = 1.05, y = drop(data$f[, 1]), label = 1:nrow(data))
ggplot(data, aes(tf = fi, shape = col, color = col)) + geom_meatballs()
ggplot(data, aes(tf = fi)) + geom_meatballs(spaghetti = FALSE) + 
  facet_wrap(~col)

# geom_lasagna is a hack of geom_line because geom_tile won't accept the 
# id-factor as a vertical axis -- adjust line width ("size") to get a proper 
# lasagna without gaps just like nonna used to make back in the old country:
ggplot(data, aes(tf = fb, order = col)) + geom_lasagna(size = 8)
ggplot(data, aes(tf = fb, order = col)) + geom_lasagna(size = 12)

# use the order-aesthetic to define the vertical ordering of the functions 
# (low values - low positions)
ggplot(data, aes(tf = f, order = -depth(f))) + geom_lasagna(size = 8)
# or use order_by to define an ordering computed directly 
# on each function's evaluations:
ggplot(data, aes(tf = f)) + geom_lasagna(order_by = mean)
ggplot(data, aes(tf = f)) + geom_lasagna(order_by = min)
ggplot(data, aes(tf = f)) + geom_lasagna(order_by = function(f) f[1])
# last one same as this:
ggplot(data, aes(tf = f, order = drop(f[, 0]))) + geom_lasagna()
# combine the two: 
ggplot(data, aes(tf = f, order = col)) + geom_lasagna(order_by = function(f) f[1]) 
# .. but facetting is broken for lasagna:
ggplot(data, aes(tf = f, order = col)) + 
  geom_lasagna(order_by = function(f) f[1])  + facet_wrap(~ col)   
}
\keyword{datasets}
