% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{is_tf_ggplot}
\alias{is_tf_ggplot}
\title{Check if object is a tf_ggplot}
\usage{
is_tf_ggplot(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"tf_ggplot"}, \code{FALSE} otherwise.
}
\description{
Check if object is a tf_ggplot
}
\examples{
p <- tf_ggplot(data.frame(x = 1))
is_tf_ggplot(p)
is_tf_ggplot(ggplot2::ggplot())
}
