% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{+.tf_ggplot}
\alias{+.tf_ggplot}
\title{Add layers to tf_ggplot objects}
\usage{
\method{+}{tf_ggplot}(e1, e2)
}
\arguments{
\item{e1}{A tf_ggplot object}

\item{e2}{A ggplot2 layer, scale, theme, etc.}
}
\value{
A modified \code{tf_ggplot} object.
}
\description{
Add layers to tf_ggplot objects
}
