% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{print.tf_ggplot}
\alias{print.tf_ggplot}
\title{Print method for tf_ggplot}
\usage{
\method{print}{tf_ggplot}(x, ...)
}
\arguments{
\item{x}{A tf_ggplot object}

\item{...}{Additional arguments}
}
\value{
\code{x}, invisibly. Called for its side effect of printing the plot.
}
\description{
Print method for tf_ggplot
}
