% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgp}
\alias{rgp}
\title{Random generator for Gaussian Processes}
\usage{
rgp(n, argvals = seq(0, 1, l = 51), scale = diff(range(argvals))/10,
  cor = c("squareexp", "wiener"), nugget = scale/200)
}
\arguments{
\item{n}{how many realizations to draw}

\item{argvals}{vector of evaluation points (\code{argvals} of the return object).
Defaults to (0, 0.02, 0.04, ..., 1).}

\item{scale}{scale parameter (see Description). Defaults to the width of the
domain divided by 10.}

\item{cor}{type of correlation structure to use. Currently available:
\code{"squareexp"} or \code{"wiener"}, see Description.}

\item{nugget}{nugget effect for additional white error noise. Defaults to
\code{scale/200} (so: very little noise)}
}
\value{
an \code{feval}-vector of length \code{n}
}
\description{
The function generates \code{n} realizations of Gaussian processes , either with
squared exponential covariance \eqn{Cov(x(t), x(t')) = \exp(-(t'-t)^2)/s) + n
\delta_{t}(t')} or Wiener process covariance \eqn{Cov(x(t), x(t')) =
\min(t',t)/s + n \delta_{t}(t')} with \code{scale} parameter s and \code{nugget} effect
n.
}
