% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\alias{smooth.tfb}
\alias{smooth.tfd}
\alias{smooth.default}
\title{Simple smoothing of \code{tf} objects}
\usage{
smooth(x, ...)

\method{smooth}{tfb}(x, ...)

\method{smooth}{tfd}(x, method = c("lowess", "rollmean", "rollmedian",
  "savgol"), ...)

\method{smooth}{default}(x, ..., kind = c("3RS3R", "3RSS", "3RSR", "3R",
  "3", "S"), twiceit = FALSE, endrule = c("Tukey", "copy"),
  do.ends = FALSE)
}
\arguments{
\item{x}{a \code{tf} object containing functional data}

\item{...}{arguments for the respective \code{method}. See Details.}

\item{method}{one of "lowess" (see \code{\link[stats:lowess]{stats::lowess()}}), "rollmean",
"rollmedian" (see \code{\link[zoo:rollmean]{zoo::rollmean()}}) or "sgolay" (see \code{\link[pracma:savgol]{pracma::savgol()}})}

\item{kind}{a character string indicating the kind of smoother required;
    defaults to \code{"3RS3R"}.}

\item{twiceit}{logical, indicating if the result should be \sQuote{twiced}.
    Twicing a smoother \eqn{S(y)} means \eqn{S(y) + S(y - S(y))}, i.e.,
    adding smoothed residuals to the smoothed values.  This decreases
    bias (increasing variance).}

\item{endrule}{a character string indicating the rule for smoothing at the
    boundary.  Either \code{"Tukey"} (default) or \code{"copy"}.}

\item{do.ends}{logical, indicating if the 3-splitting of ties should
    also happen at the boundaries (ends).  This is only used for
    \code{kind = "S"}.}
}
\value{
a smoothed version of the input. For some methods/options, the
smoothed functions may be shorter than the original ones (at both ends).
}
\description{
Apply running means or medians, \code{lowess} or Savitzky-Golay
filtering to smooth functional data. This does nothing for \code{tfb}-objects,
which should be smoothed by using a smaller basis / stronger penalty.\cr
The \code{\link[stats:smooth]{stats::smooth()}} function is re-exported as the default method.
}
\details{
\code{smooth.tfd} overrides/automatically sets some defaults of the used
methods:
\itemize{
\item \strong{\code{lowess}} uses a span parameter of \code{f} = .15 (instead of .75)
by default.
\item \strong{\code{rollmean}/\code{median}} use a window size of \code{k} = <number of
  grid points>/20 (i.e., the nearest odd integer to that) and sets \code{fill= "extend"} (i.e., constant extrapolation to replace missing values at the
extremes of the domain) by default. Use \code{fill= NA} for \code{zoo}'s default
behavior of shortening the smoothed series.
\item \strong{\code{savgol}} uses a window size of \code{k} = <number of
  grid points>/10 (i.e., the nearest odd integer to that).
}
}
\examples{
f <- tf_sparsify(tf_jiggle(tf_rgp(4, 201L, nugget = .05)))
f_lowess <- smooth(f, "lowess")
# these methods ignore the distances between arg-values:
f_mean <- smooth(f, "rollmean")
f_median <- smooth(f, "rollmean", k = 31)
f_sg <- smooth(f, "savgol", fl = 31)
layout(t(1:4))
plot(f, points = FALSE)
plot(f_lowess, points = FALSE)
lines(smooth(f, "lowess", f = .9), col = 2, alpha= .2)
plot(f_mean, points = FALSE)
lines(f_median, col = 2, alpha= .2) # note constant extrapolation
plot(f_sg, points = FALSE) 
}
