% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{tf_depth}
\alias{tf_depth}
\alias{tf_depth.matrix}
\alias{tf_depth.tf}
\title{Functional Data Depth}
\usage{
tf_depth(x, depth = "MBD", na.rm = TRUE, ...)

\method{tf_depth}{matrix}(x, depth = "MBD", na.rm = TRUE,
  arg = unlist(find_arg(x, NULL)), ...)

\method{tf_depth}{tf}(x, depth = "MBD", na.rm = TRUE, arg = NULL,
  ...)
}
\arguments{
\item{x}{\code{tf} (or a matrix of evaluations)}

\item{depth}{currently available: "MBD", i.e. modified band depth}

\item{na.rm}{TRUE remove missing observations?}

\item{...}{further arguments handed to the function computing the respective tf_depth.}

\item{arg}{grid of evaluation points}
}
\value{
vector of tf_depth values
}
\description{
Data tf_depth for functional data.
Currently implemented: Modified Band-2 Depth, see reference.
}
\references{
Sun, Y., Genton, M. G., & Nychka, D. W. (2012).
Exact fast computation of band tf_depth for large functional datasets:
How quickly can one million curves be ranked?. \emph{Stat}, \strong{1}(1), 68-74.
Lopez-Pintado, S. and Romo, J. (2009).
On the Concept of Depth for Functional Data.
\emph{Journal of the American Statistical Association}, \strong{104}, 718-734.
}
