% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{tf_jiggle}
\alias{tf_jiggle}
\alias{tf_sparsify}
\title{Make a \code{tf} (more) irregular}
\usage{
tf_jiggle(f, ...)

tf_sparsify(f, dropout = 0.5, ...)
}
\arguments{
\item{f}{a \code{tfd} object}

\item{...}{not used currently}

\item{dropout}{how many values of \code{f} to drop, defaults to 50\\%.}
}
\description{
Randomly create some irregular functional data from regular ones.\cr
\strong{jiggle} it by randomly moving around its \code{arg}-values. Only for \code{tfd}.\cr
\strong{sparsify} it by setting (100*\code{dropout})\\% of its values to \code{NA}.
}
