% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{tf_zoom}
\alias{tf_zoom}
\alias{tf_zoom.tfd}
\alias{tf_zoom.tfb}
\title{Functions to zoom in/out on functions}
\usage{
tf_zoom(f, begin, end, ...)

\method{tf_zoom}{tfd}(f, begin = domain(f)[1], end = domain(f)[2], ...)

\method{tf_zoom}{tfb}(f, begin = domain(f)[1], end = domain(f)[2], ...)
}
\arguments{
\item{f}{a \code{tf}-object}

\item{begin}{numeric vector of length 1 or \code{length(f)}.
Defaults to the lower limit of the domain of \code{f}.}

\item{end}{numeric vector of length 1 or \code{length(f)}.
Defaults to the upper limit of the domain of \code{f}.}

\item{...}{not used}
}
\value{
an object like \code{f} on a new domain (potentially).
Note that regular functional data and functions in basis representation will
be turned into irregular \code{tfd}-objects iff \code{begin} or \code{end} are not scalar.
}
\description{
These are used to redefine or restrict the \code{domain} of \code{tf} objects.
}
\examples{
  (x <- tf_rgp(10))
  plot(x)
  tf_zoom(x, .5, .9)
  lines(tf_zoom(x, .5, .9), col = "red")
  points(tf_zoom(x, seq(0, .5, l = 10), seq(.5, 1, l = 10)), col = "blue")
}
