% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R, R/tfb-class.R
\name{as.tfb}
\alias{as.tfb}
\alias{as.data.frame.tfb}
\alias{as.matrix.tfb}
\alias{tfb}
\alias{tfb_wavelet}
\title{Constructors & convertors for functional data in basis representation}
\usage{
as.tfb(data, basis = c("spline", "fpc"), ...)

\method{as.data.frame}{tfb}(x, row.names = NULL, optional = FALSE,
  arg = NULL, ...)

\method{as.matrix}{tfb}(x, arg = NULL, ...)

tfb(data, basis = c("spline", "fpc", "wavelet"), ...)

tfb_wavelet(data, ...)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} or \code{list} of suitable shape, or another
\code{tf}-object containing functional data.}

\item{basis}{either "\code{spline}" (see \code{\link[=tfb_spline]{tfb_spline()}}, the default) or "\code{fpc}" (see \code{\link[=tfb_fpc]{tfb_fpc()}}).
(\code{wavelet} not implemented yet)}

\item{...}{further arguments for \code{\link[=tfb_spline]{tfb_spline()}} or \code{\link[=tfb_fpc]{tfb_fpc()}}}

\item{x}{an \code{tfb} object}

\item{row.names}{not used}

\item{optional}{not used}

\item{arg}{optional vector of argument values}
}
\value{
a \code{tfb}-object (or a \code{data.frame}/\code{matrix} for the conversion
functions, obviously.)
}
\description{
Various constructor and conversion methods.
}
\details{
\code{tfb} is a wrapper for functions that set up spline-, principal component-
or wavelet-based representations of functional data. For all three, the
input data $x_i(t)$ are represented as weighted sums of a set of common basis
functions $B_k(t); k = 1,\dots, K$ identical for all
observations and weight or coefficient vectors $b_i = (b_{i1}, \dots, b_{iK})$
estimated for each observation: $x_i(t) = B_1(t) b_{i1} + \dots + B_K(t) b_{iK}$.
Depending on the value of \code{basis}, the basis functions $B(t)$ will either be
\code{spline`` functions or the first few estimated eigenfunctions of the covariance operator of the $x(t)$ (}fpc\code{) or wavelets (}wavelet`).

See \code{\link[=tfb_spline]{tfb_spline()}} for spline based basis representation (the default).
See \code{\link[=tfb_fpc]{tfb_fpc()}} for using an FPC representation with an orthonormal basis estimated from the
data instead.
}
\seealso{
\code{\link[=tfb_fpc]{tfb_fpc()}}, \code{\link[=tfb_spline]{tfb_spline()}}
}
