% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pillar.R
\name{type_sum.tf}
\alias{type_sum.tf}
\alias{obj_sum.tf}
\alias{pillar_shaft.tf}
\title{Format tidy functional data for tibbles}
\usage{
\method{type_sum}{tf}(x, ...)

\method{obj_sum}{tf}(x)

\method{pillar_shaft}{tf}(x, ...)
}
\arguments{
\item{x}{object of class tf}

\item{...}{ignored}
}
\description{
Summarize tidy functional data for tibble
}
\details{
see \code{\link[pillar:type_sum]{pillar::type_sum()}}
}
