% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{funplot}
\alias{funplot}
\alias{plot.tf}
\alias{lines.tf}
\alias{points.tf}
\title{Visualization functions for \code{tf}s}
\usage{
funplot(f, argvals, n_grid = 50, points = is_irreg(f),
  type = c("spaghetti", "lasagna"), alpha = min(1, max(0.05, 2/length(f))))

\method{plot}{tf}(x, y, n_grid = 50, points = is_irreg(x),
  type = c("spaghetti", "lasagna"), alpha = min(1, max(0.05, 2/length(x))),
  ...)

\method{lines}{tf}(x, argvals, n_grid = 50, alpha = min(1, max(0.05,
  2/length(x))), ...)

\method{points}{tf}(x, argvals, n_grid = NA, alpha = min(1, max(0.05,
  2/length(x))), interpolate = FALSE, ...)
}
\arguments{
\item{f}{an \code{tf} object}

\item{argvals}{(optional) numeric vector to be used as \code{argvals}. See details.}

\item{n_grid}{minimal size of equidistant grid used for plotting,
defaults to 50. See details.}

\item{points}{should the original evaluation points be marked by points?
Defaults to \code{TRUE} for irregular \code{feval} and FALSE for all others}

\item{type}{"spaghetti": line plots, "lasagna": heat maps.}

\item{alpha}{\href{grDevices::rgb()}{alpha-value} for noodle transparency.
Defaults to 2/(no. of observations). Lower is more transparent.}

\item{x}{an \code{tf} object}

\item{y}{(optional) numeric vector to be used as \code{argvals} (i.e., for the \strong{x}-axis...!)}

\item{...}{additional arguments for \code{\link[=matplot]{matplot()}} ("spaghetti") or \code{\link[=image]{image()}} ("lasagna")}

\item{interpolate}{should functions be evaluated (i.e., inter-/extrapolated)
for argvals for which no original data is available? Only relevant for
feval, defaults to FALSE}
}
\value{
for \code{funplot}: the \code{ggplot}-object for further modification. For the
others: the plotted \code{tf}-object, invisibly.
}
\description{
Some \code{base} and \code{ggplot}-functions for displaying functional data in
spaghetti- (i.e., line plots) and lasagna- (i.e., heat map) flavors.
}
\details{
\code{funplot} uses \code{ggplot2}, the others use the \code{base}-system.

If no \code{argvals} are provided, evaluation points (\code{argvals}) for the functions
are given by the union of the \code{tf}'s \code{argvals} and an equidistant grid
over its domain with \code{n_grid} points. If you want to only see the original
data for \code{feval}-objects without inter-/extrapolation, use \code{n_grid < 1} or
\code{n_grid = NA}.
}
\references{
Swihart, B. J., Caffo, B., James, B. D., Strand, M., Schwartz, B. S., & Punjabi, N. M. (2010).
Lasagna plots: a saucy alternative to spaghetti plots. \emph{Epidemiology (Cambridge, Mass.)}, \strong{21}(5), 621-625.
}
