% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfun-package.R
\docType{package}
\name{tidyfun-package}
\alias{tidyfun}
\alias{tidyfun-package}
\title{tidyfun: Tidy Functional Data Wrangling and Visualization}
\description{
\code{tidyfun} makes data wrangling and exploratory analysis for functional data
in \code{R} easier.\cr\cr
The \code{tf} package defines the core functional data classes and methods.
\code{tidyfun} builds on those objects with:
\itemize{
\item data-frame workflows for functional data: \code{\link[=tf_gather]{tf_gather()}} etc.
\item graphics helpers for \code{tf} objects and tidy functional data frames: \code{\link[=geom_spaghetti]{geom_spaghetti()}}, \code{\link[=gglasagna]{gglasagna()}}
\item documentation and examples that show how to use \code{tf} objects in tidy workflows
\cr\cr
\cr
Start with \code{vignette(package = "tidyfun")} for the article index.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyfun/tidyfun}
  \item \url{https://tidyfun.github.io/tidyfun/}
  \item Report bugs at \url{https://github.com/tidyfun/tidyfun/issues}
}

}
\author{
\strong{Maintainer}: Fabian Scheipl \email{fabian.scheipl@googlemail.com} (\href{https://orcid.org/0000-0001-8172-3603}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Fabian Scheipl \email{fabian.scheipl@googlemail.com} (\href{https://orcid.org/0000-0001-8172-3603}{ORCID}) [copyright holder]
  \item Jeff Goldsmith
  \item Julia Wrobel (\href{https://orcid.org/0000-0001-6783-1421}{ORCID})
}

Other contributors:
\itemize{
  \item Maximilian Mücke (\href{https://orcid.org/0009-0000-9432-9795}{ORCID}) [contributor]
}

}
\keyword{internal}
