% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfun-package.r
\docType{package}
\name{tidyfun}
\alias{tidyfun}
\alias{tidyfun-package}
\title{tidyfun}
\description{
\code{tidyfun} makes data wrangling and exploratory analysis for functional data
in \code{R} easier.\cr\cr
\code{tidyfun} provides:
\itemize{
\item new data types for representing functional data: \code{\link[=tfd]{tfd()}} & \code{\link[=tfb]{tfb()}}
\item arithmetic operators (\code{\link[=Ops.tf]{Ops.tf()}}),
\item descriptive statistics: e.g. \code{\link[=mean.tf]{mean.tf()}}, \code{\link[=median.tf]{median.tf()}}
\item and graphics functions for such data: \code{\link[=plot.tf]{plot.tf()}}, \code{\link[=geom_spaghetti]{geom_spaghetti()}}, \code{\link[=gglasagna]{gglasagna()}}
\item functions to do smoothing (\code{\link[=tf_smooth.tfd]{tf_smooth.tfd()}}), differentiation and integration (\code{\link[=tf_derive.tfd]{tf_derive.tfd()}})
\item \code{tidyverse}-verbs for handling functional data inside data frames: \code{\link[=tf_gather]{tf_gather()}} etc.
\cr\cr
\cr
Also see \code{vignette("Introducing tidyfun", "tidyfun")} for a brief introduction.
}
}
