% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{translate_old_tf_layer}
\alias{translate_old_tf_layer}
\title{Translate a geom_spaghetti/geom_meatballs/geom_errorband layer for tf_ggplot}
\usage{
translate_old_tf_layer(layer, e1)
}
\arguments{
\item{layer}{A LayerInstance using StatTf or StatErrorband}

\item{e1}{The tf_ggplot object (may be modified to clean up plot-level y)}
}
\value{
A list with \code{plot_obj} (possibly modified e1) and \code{layers} (list of
new layers to add)
}
\description{
Remaps old-style y/ymin/ymax tf aesthetics to tf/tf_ymin/tf_ymax and
substitutes the appropriate standard ggplot2 geom.
}
\keyword{internal}
