% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-vctrs.R
\name{vec_cast.tfd_reg}
\alias{vec_cast.tfd_reg}
\alias{vec_cast.tfd_irreg}
\alias{vec_cast.tfd_reg.tfd_reg}
\alias{vec_cast.tfd_reg.tfd_irreg}
\alias{vec_cast.tfd_irreg.tfd_reg}
\alias{vec_cast.tfd_irreg.tfd_irreg}
\alias{vctrs}
\alias{vec_ptype2.tfd_reg}
\alias{vec_ptype2.tfd_reg.tfd_reg}
\alias{vec_ptype2.tfd_reg.tfd_irreg}
\alias{vec_ptype2.tfd_irreg}
\alias{vec_ptype2.tfd_irreg.tfd_reg}
\alias{vec_ptype2.tfd_irreg.tfd_irreg}
\alias{vec_ptype2_tfd_tfd}
\alias{vec_cast.tfb_spline}
\alias{vec_cast.tfb_fpc}
\alias{vec_cast.tfb_spline.tfb_spline}
\alias{vec_cast.tfb_spline.tfb_fpc}
\alias{vec_cast.tfb_fpc.tfb_spline}
\alias{vec_cast.tfb_fpc.tfb_fpc}
\alias{vec_ptype2.tfb_spline}
\alias{vec_ptype2.tfb_spline.tfb_spline}
\alias{vec_ptype2.tfb_spline.tfb_fpc}
\alias{vec_ptype2.tfb_fpc}
\alias{vec_ptype2.tfb_fpc.tfb_spline}
\alias{vec_ptype2.tfb_fpc.tfb_fpc}
\alias{vec_ptype2_tfb_tfb}
\title{vctrs methods for \code{tf} objects}
\usage{
\method{vec_cast}{tfd_reg}(x, to, ...)

\method{vec_cast}{tfd_irreg}(x, to, ...)

\method{vec_cast.tfd_reg}{tfd_reg}(x, to, ...)

\method{vec_cast.tfd_reg}{tfd_irreg}(x, to, ...)

\method{vec_cast.tfd_irreg}{tfd_reg}(x, to, ...)

\method{vec_cast.tfd_irreg}{tfd_irreg}(x, to, ...)

\method{vec_ptype2}{tfd_reg}(x, y, ...)

\method{vec_ptype2.tfd_reg}{tfd_reg}(x, y, ...)

\method{vec_ptype2.tfd_reg}{tfd_irreg}(x, y, ...)

\method{vec_ptype2}{tfd_irreg}(x, y, ...)

\method{vec_ptype2.tfd_irreg}{tfd_reg}(x, y, ...)

\method{vec_ptype2.tfd_irreg}{tfd_irreg}(x, y, ...)

vec_ptype2_tfd_tfd(x, y, ...)

\method{vec_cast}{tfb_spline}(x, to, ...)

\method{vec_cast}{tfb_fpc}(x, to, ...)

\method{vec_cast.tfb_spline}{tfb_spline}(x, to, ...)

\method{vec_cast.tfb_spline}{tfb_fpc}(x, to, ...)

\method{vec_cast.tfb_fpc}{tfb_spline}(x, to, ...)

\method{vec_cast.tfd_irreg}{tfd_irreg}(x, to, ...)

\method{vec_ptype2}{tfb_spline}(x, y, ...)

\method{vec_ptype2.tfb_spline}{tfb_spline}(x, y, ...)

\method{vec_ptype2.tfb_spline}{tfb_fpc}(x, y, ...)

\method{vec_ptype2}{tfb_fpc}(x, y, ...)

\method{vec_ptype2.tfb_fpc}{tfb_spline}(x, y, ...)

\method{vec_ptype2.tfb_fpc}{tfb_fpc}(x, y, ...)

vec_ptype2_tfb_tfb(x, y, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}

\item{y}{Vector types.}
}
\description{
These functions are the extensions that allow \code{tidyfun} vectors
to work with \code{vctrs}.

These functions are the extensions that allow \code{tidyfun} vectors
to work with \code{vctrs}.
}
