% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{add_tf_aes_to_mapping}
\alias{add_tf_aes_to_mapping}
\title{Add a tf aesthetic to an ggplot2 mapping object}
\usage{
add_tf_aes_to_mapping(mapping, aes_name, val_col, arg_col, id_col)
}
\description{
Add a tf aesthetic to an ggplot2 mapping object
}
\keyword{internal}
