% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chf_df}
\alias{chf_df}
\title{Congestive heart failure accelerometry data}
\description{
Activity data from a study of congestive heart failure.
Data were originally presented
in "Multilevel Matrix-Variate Analysis and its Application
to Accelerometry-Measured Physical Activity in Clinical
Populations" by Huang et al.; these data are public,
with download information in the paper.
}
\keyword{data}
