% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dti_df}
\alias{dti_df}
\title{Diffusion tensor imaging data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 382 rows and 6 columns.
}
\usage{
dti_df
}
\description{
Fractional anisotropy (FA) tract profiles for the corpus
callosum (cca) and the right corticospinal tract (rcst).
Accompanying the tract profiles are the subject ID numbers,
visit number, total number of scans, multiple sclerosis case
status and Paced Auditory Serial Addition Test (pasat) score.

Data are also included in the \CRANpkg{refund} package in another
format.
}
\details{
If you use this data as an example in written work, please
include the following acknowledgment: "The MRI/DTI data were
collected at Johns Hopkins University and the Kennedy-Krieger
Institute"
}
\examples{
dti_df
}
\seealso{
Other tidyfun datasets: 
\code{\link{chf_df}}
}
\concept{tidyfun datasets}
\keyword{data}
