% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.feval}
\alias{evaluate.fbase}
\alias{evaluate.data.frame}
\title{Evaluate \code{fvector}s, both inside or outside a \code{data.frame}}
\usage{
evaluate(object, argvals, ...)

\method{evaluate}{feval}(object, argvals, ...)

\method{evaluate}{fbase}(object, argvals, ...)

\method{evaluate}{data.frame}(object, argvals, ...)
}
\arguments{
\item{object}{an \code{fvector} or a \code{data.frame}-like object with \code{fvector} columns}

\item{argvals}{optional evaluation grid, defaults to \code{argvals(object)}.}

\item{...}{optional: names of the \code{fvector}-columns to unnest}
}
\description{
The \code{evaluate.data.frame} method evaluates \code{fvector}-columns inside a \code{data.frame}
into list columns of smaller \code{data.frames} containing the functions' argvals and
evaluations. Its \code{argvals}-argument can be a list of \code{argvals}-vector
used as the \code{argvals} argument for the \code{\link[=evaluate]{evaluate()}}-method for the respective
\code{fvector}-columns in \code{object}.
}
\seealso{
\code{?`[.fvector`}
}
