% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feval-def.R, R/fvector-conversion.R
\name{feval}
\alias{feval}
\alias{feval.matrix}
\alias{feval.data.frame}
\alias{feval.list}
\alias{feval.fvector}
\alias{as.feval}
\alias{as.feval.matrix}
\alias{as.feval.data.frame}
\alias{as.feval.list}
\alias{as.feval.fbase}
\alias{as.feval_irreg}
\alias{as.data.frame.feval}
\alias{as.matrix.feval}
\title{Constructors for functional data evaluated on grids of argument values}
\usage{
feval(data, ...)

\method{feval}{matrix}(data, argvals = NULL, domain = NULL,
  evaluator = approx_linear, signif = 4, ...)

\method{feval}{data.frame}(data, id = 1, argvals = 2, value = 3,
  domain = NULL, evaluator = approx_linear, signif = 4, ...)

\method{feval}{list}(data, argvals = NULL, domain = NULL,
  evaluator = approx_linear, signif = 4, ...)

\method{feval}{fvector}(data, argvals = NULL, domain = NULL,
  evaluator = approx_linear, signif = 4, ...)

as.feval(data, ...)

\method{as.feval}{matrix}(data, argvals = NULL, domain = NULL, ...)

\method{as.feval}{data.frame}(data, id = 1, argvals = 2, value = 3,
  domain = NULL, ...)

\method{as.feval}{list}(data, argvals = NULL, domain = NULL, ...)

\method{as.feval}{fbase}(data, argvals = NULL, domain = NULL, ...)

as.feval_irreg(data, signif = NULL, ...)

\method{as.data.frame}{feval}(x, rownames = NULL, optional = FALSE,
  argvals = NULL, interpolate = FALSE, ...)

\method{as.matrix}{feval}(x, argvals = NULL, interpolate = FALSE, ...)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} or \code{list} of suitable shape, or another \code{fvector}-object.}

\item{...}{not used, except for \code{feval.fvector} -- specify \code{argvals} and \code{ìnterpolate = TRUE} to
turn an irregular \code{feval} into a regular one, see examples...}

\item{argvals}{\code{numeric}, or list of \code{numeric}s. The evaluation grid. See Details.}

\item{domain}{range of the \code{argvals}.}

\item{evaluator}{a function accepting arguments \code{x, argvals, evaluations}. See details.}

\item{signif}{significant digits of the "resolution" of the evaluation grid.  See details.}

\item{id}{The name/number of the column defining which data belong to which function.}

\item{value}{The name/number of the column containing the function evaluations.}

\item{x}{an \code{feval} object}

\item{rownames}{not used}

\item{optional}{not used}

\item{interpolate}{should functions be evaluated (i.e., inter-/extrapolated)
for \code{argvals} for which no original data is available? Only relevant for
\code{feval}, defaults to TRUE.}
}
\value{
an \code{feval}-object (or a \code{data.frame}/\code{matrix} for the conversion functions, obviously.)
}
\description{
Various constructor and conversion methods.
}
\details{
\strong{\code{evaluator}}: must be a \code{function(x, argvals, evaluations)} that returns
the function's (approximated/interpolated) values at locations \code{x} based on
the \code{evaluations} available at locations \code{argvals}.
Available \code{evaluator}-functions:
\itemize{
\item \code{approx_linear} for linear interpolation without extrapolation (i.e.,
\code{\link[zoo:na.approx]{zoo::na.approx()}} with \code{na.rm = FALSE}),
\item \code{approx_spline} for cubic spline interpolation, (i.e., \code{\link[zoo:na.spline]{zoo::na.spline()}}
with \code{na.rm = FALSE}),
\item \code{approx_fill_extend} for linear interpolation and constant extrapolation
(i.e., \code{\link[zoo:na.fill]{zoo::na.fill()}} with \code{fill = "extend"})
\item \code{approx_locf} for "last observation carried forward"  (i.e.,
\code{\link[zoo:na.locf]{zoo::na.locf()}} with \code{na.rm = FALSE} and
\item \code{approx_nocb} for "next observation carried backward" (i.e.,
\code{\link[zoo:na.locf]{zoo::na.locf()}} with \code{na.rm = FALSE, fromLast = TRUE}).
}

See \code{tidyfun:::zoo_wrapper} and \code{tidyfun:::approx_linear}, which is simply
\code{zoo_wrapper(zoo::na.approx, na.rm = FALSE)}, for examples of implementations of
this.

\strong{\code{signif}}: \code{argvals} that are equivalent to this significant digit are
treated as identical. E.g., if an evaluation of f(t) is available at t=1 and a function
value is requested at t = 1.001, f(1) will be returned if \code{signif} < 4.
}
\examples{
#turn irregular to regular feval
#TODO
}
