% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{argvals}
\alias{argvals}
\alias{evaluations}
\alias{n_evaluations}
\alias{domain}
\alias{evaluator<-}
\alias{argvals<-}
\alias{coef.fbase}
\alias{rev.fvector}
\alias{mean.fvector}
\alias{median.fvector}
\alias{quantile.fvector}
\alias{sd}
\alias{sd.default}
\alias{sd.fvector}
\alias{var}
\alias{var.default}
\alias{var.fvector}
\title{Utility functions for \code{fvector}-objects}
\usage{
argvals(f)

evaluations(f)

n_evaluations(f)

domain(f)

evaluator(x) <- value

argvals(x) <- value

\method{coef}{fbase}(object, ...)

\method{rev}{fvector}(x)

\method{mean}{fvector}(x, ...)

\method{median}{fvector}(x, na.rm = FALSE, ...)

\method{quantile}{fvector}(x, probs = seq(0, 1, 0.25), na.rm = FALSE,
  names = TRUE, type = 7, ...)

sd(x, na.rm = FALSE)

\method{sd}{default}(x, na.rm = FALSE)

\method{sd}{fvector}(x, na.rm = FALSE)

var(x, y = NULL, na.rm = FALSE, use)

\method{var}{default}(x, y = NULL, na.rm = FALSE, use)

\method{var}{fvector}(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{f}{an \code{fvector} object}

\item{x}{numeric vector whose sample quantiles are wanted, or an
    object of a class for which a method has been defined (see also
    \sQuote{details}). \code{\link{NA}} and \code{NaN} values are not
    allowed in numeric vectors unless \code{na.rm} is \code{TRUE}.}

\item{value}{for \code{evaluator<-}: a function that can be used to interpolate an \code{feval}. Needs to
accept vector arguments \code{x}, \code{argvals}, \code{evaluations} and return
evaluations of the function defined by \code{argvals}, \code{evaluations} at \code{x}
for \code{argvals<-}: a list of grid points, for internal use only.}

\item{object}{as usual}

\item{...}{dots}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}

\item{names}{logical; if true, the result has a \code{\link{names}}
    attribute.  Set to \code{FALSE} for speedup with many \code{probs}.}

\item{type}{an integer between 1 and 9 selecting one of the
    nine quantile algorithms detailed below to be used.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with
    compatible dimensions to \code{x}.   The default is equivalent to
    \code{y = x} (but more efficient).}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}
}
\description{
A bunch of methods that do what they say.
}
