% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feval-methods.R
\name{argvals}
\alias{argvals}
\alias{evaluations}
\alias{n_evaluations}
\alias{evaluator<-}
\alias{range.fvector}
\title{Utility functions for `fvector`-objects}
\usage{
argvals(f)

evaluations(f)

n_evaluations(f)

evaluator(x) <- value

\method{range}{fvector}(x, na.rm = FALSE)
}
\arguments{
\item{f}{an `fvector` object}
}
\description{
A bunch of methods that do what they say.
}
