% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvector-methods.R
\name{argvals}
\alias{argvals}
\alias{evaluations}
\alias{n_evaluations}
\alias{domain}
\alias{evaluator<-}
\alias{range.fvector}
\title{Utility functions for \code{fvector}-objects}
\usage{
argvals(f)

evaluations(f)

n_evaluations(f)

domain(f)

evaluator(x) <- value

\method{range}{fvector}(..., na.rm = FALSE)
}
\arguments{
\item{f}{an \code{fvector} object}

\item{value}{a function that can be used to interpolate an \code{feval}. Needs to
accept vector arguments \code{x}, \code{argvals}, \code{evaluations} and return
evaluations of the function defined by \code{argvals}, \code{evaluations} at \code{x}}

\item{...}{dots}

\item{na.rm}{as usual}
}
\description{
A bunch of methods that do what they say.
}
