% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-capellini.R
\docType{data}
\name{ggcapellini}
\alias{ggcapellini}
\alias{StatCapellini}
\alias{stat_capellini}
\alias{geom_capellini}
\alias{geom_cappellini}
\alias{geom_capelini}
\alias{geom_cappelini}
\alias{GeomCapellini}
\title{Glyph plots for \code{tf} objects}
\usage{
stat_capellini(
  mapping = NULL,
  data = NULL,
  geom = "capellini",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  arg = NULL,
  add_lines = FALSE,
  add_boxes = TRUE,
  width = NULL,
  height = NULL,
  ...
)

geom_capellini(
  mapping = NULL,
  data = NULL,
  stat = "capellini",
  position = "identity",
  ...,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  arg = NULL,
  add_lines = TRUE,
  add_boxes = TRUE,
  width = NULL,
  height = NULL,
  box.colour = "#0000001A",
  box.linetype = 1,
  box.fill = NA,
  box.size = 0.1,
  box.alpha = 0.1,
  line.colour = "black",
  line.linetype = 2,
  line.size = 0.3,
  line.alpha = 0.5
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data, either as a
\code{ggproto} \code{Geom} subclass or as a string naming the geom stripped of the
\code{geom_} prefix (e.g. \code{"point"} rather than \code{"geom_point"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{na.rm}{remove NAs? defaults to \code{TRUE}}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{arg}{where to evaluate \code{tf} -- defaults to the default ;)}

\item{add_lines}{should a reference line in the middle of the range of the
functions' values be added to each glyph? defaults to TRUE}

\item{add_boxes}{should a box be added to frame each glyph? defaults to TRUE}

\item{width}{the width of the glyphs. Defaults to 2/3 of the \code{\link[ggplot2:resolution]{ggplot2::resolution()}}
of the variable for the \code{x}-aesthetic, this will be too small if any values
are close together.}

\item{height}{the height of the glyphs. Defaults to 2/3 of the \code{\link[ggplot2:resolution]{ggplot2::resolution()}}
of the variable for the \code{y}-aesthetic, this will be too small if any values
are close together.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{that's "capellini"!}

\item{box.colour}{aesthetic property of the box}

\item{box.linetype}{aesthetic property of the box}

\item{box.fill}{aesthetic property of the box}

\item{box.size}{aesthetic property of the box}

\item{box.alpha}{aesthetic property of the box}

\item{line.colour}{aesthetic property of the reference line}

\item{line.linetype}{aesthetic property of the reference line}

\item{line.size}{aesthetic property of of the reference line}

\item{line.alpha}{aesthetic property of the reference line}
}
\description{
Plots a miniature glyph or sparkline for each entry of a \code{tf}-object.
(Capellini are tiny spaghetti -- \emph{angel hair} pasta.) Aesthetics \code{x} and \code{y}
specify the location of the glyphs, the \code{tf} aesthetic defines their shapes.
To accommodate all my fellow idiots, \code{geom_cappelini},  \code{geom_cappellini} and
\code{geom_capelini} also work.
}
\section{\code{tf} aesthetic}{

Mandatory. Used to designate a column of class \code{tf} to be visualized as glyphs.
}

\examples{
\dontrun{ # takes a little too long for CRAN
library(ggplot2)
library(tidyverse)
weather <- fda::CanadianWeather
canada <- data.frame(
  place = weather$place,
  region = weather$region,
  lat = weather$coordinates[,1],
  lon = -weather$coordinates[,2],
  region = weather$region)
canada$temp <- tfd(t(weather$dailyAv[,,1]), arg = 1:365)
canada$precipl10 <- tfd(t(weather$dailyAv[,,3]), arg = 1:365) \%>\% tf_smooth
canada_map <-
  data.frame(maps::map("world", "Canada", plot = FALSE)[c("x", "y")])
# map of canada with annual temperature averages in red, precipitation in blue:
ggplot(canada, aes(x = lon, y = lat)) +
  geom_capellini(aes(tf = precipl10), width = 3, height = 5, colour = "blue") +
  geom_capellini(aes(tf = temp), width = 3, height = 5, colour = "red") +
  geom_path(data = canada_map, aes(x = x, y = y), alpha = .1) +
  coord_quickmap()

ggplot(canada, aes(x = lon, y = lat, colour = region)) +
  geom_capellini(aes(tf = precipl10), width = 5, height = 3,
    line.linetype = 1, box.fill = "white", box.alpha=.5, box.colour = NA)
}
}
\seealso{
GGally::glyphs

Other tidyfun visualization: 
\code{\link{gglasagna}()},
\code{\link{ggspaghetti}}

Other tidyfun visualization: 
\code{\link{gglasagna}()},
\code{\link{ggspaghetti}}

Other tidyfun visualization: 
\code{\link{gglasagna}()},
\code{\link{ggspaghetti}}

Other tidyfun visualization: 
\code{\link{gglasagna}()},
\code{\link{ggspaghetti}}

Other tidyfun visualization: 
\code{\link{gglasagna}()},
\code{\link{ggspaghetti}}

Other tidyfun visualization: 
\code{\link{gglasagna}()},
\code{\link{ggspaghetti}}

Other tidyfun visualization: 
\code{\link{gglasagna}()},
\code{\link{ggspaghetti}}

Other tidyfun visualization: 
\code{\link{gglasagna}()},
\code{\link{ggspaghetti}}
}
\concept{tidyfun visualization}
\keyword{datasets}
