% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglasagna.R
\name{gglasagna}
\alias{gglasagna}
\title{Lasagna plots for \code{tf}s using \code{ggplot2}}
\usage{
gglasagna(data, y, order = NULL, label = NULL, arg = NULL,
  order_by = NULL, order_ticks = TRUE, order_ticks_color = "black",
  order_ticks_linetype = 2, order_ticks_alpha = 0.5)
}
\arguments{
\item{data}{the data. Duhh.}

\item{y}{bare name of the \code{tf} column to visualize}

\item{order}{(optional) bare name of a column in \code{data} to define vertical
order of lasagna layers.}

\item{label}{(optional) bare name of a column in \code{data} to define labels for
lasagna layers. Defaults to names of \code{y}, if present, or row numbers.}

\item{arg}{\code{arg} to evaluate \code{y} om}

\item{order_by}{a function applied to each row in \code{y[, arg]} that must
return a scalar value to define the order of lasagna layers.}

\item{order_ticks}{add horizontal lines indicating borders between levels of
\code{order} (if discrete)? Defaults to true.}

\item{order_ticks_color}{defaults to "black"}

\item{order_ticks_linetype}{defaults to 2}

\item{order_ticks_alpha}{defaults to 0.5}
}
\value{
a `ggplot2`` object
}
\description{
Lasagna plots show one color bar for each function.
}
\details{
The vertical order of the lasagna layers is \strong{increasing} in
\itemize{
\item \code{order} (if provided),
\item the values returned by \code{order_by} (if provided),
\item and the row number of the observations.
}

i.e., lowest values are on top so that by default the first layer is the first
observation in \code{data} and the vertical order of the layers corresponds to the
ordering of observations for \code{arrange(data, order, order_by(value), row_number())}.
}
\examples{
set.seed(1221)
data = expand.grid(group = 1:5, rep = 1:10) 
data = dplyr::mutate(data, id = paste(group, rep, sep = "-"))
data$f = rgp(50)
data$fi = sparsify(data$f)
data$fb = tfb(data$f)
gglasagna(data, f, label = id)
gglasagna(data, f, label = id, order = group)
gglasagna(data, f, label = id, order = group)
gglasagna(data, f, label = id, order = depth(f))
gglasagna(data, f, label = id, order_by = first) + 
  facet_wrap(~group, scales = "free")
#FIXME: render errors for weird arg lenght (e.g. 93) 
}
