% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{ggplot_build.tf_ggplot}
\alias{ggplot_build.tf_ggplot}
\title{ggplot_build method for tf_ggplot}
\usage{
\method{ggplot_build}{tf_ggplot}(plot, ...)
}
\arguments{
\item{plot}{A tf_ggplot object}

\item{...}{Additional arguments passed through from \code{\link[ggplot2:ggplot_build]{ggplot2::ggplot_build()}}.}
}
\value{
A built ggplot object (class \code{ggplot_built}).
}
\description{
ggplot_build method for tf_ggplot
}
