% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spaghetti.R
\docType{data}
\name{ggspaghetti}
\alias{ggspaghetti}
\alias{StatTf}
\alias{stat_tf}
\alias{geom_spaghetti}
\alias{GeomSpaghetti}
\alias{geom_meatballs}
\alias{GeomMeatballs}
\title{Spaghetti plots for \code{tf} objects}
\usage{
stat_tf(mapping = NULL, data = NULL, geom = "spaghetti",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, arg = NULL, ...)

geom_spaghetti(mapping = NULL, data = NULL, position = "identity",
  na.rm = TRUE, show.legend = NA, inherit.aes = TRUE, arg = NULL,
  ...)

geom_meatballs(mapping = NULL, data = NULL, position = "identity",
  na.rm = TRUE, show.legend = NA, inherit.aes = TRUE, arg = NULL,
  spaghetti = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{remove NAs? defaults to \code{TRUE}}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{arg}{where to evaluate \code{tf} -- defaults to the default ;)}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{spaghetti}{plot noodles along with meatballs? defaults to true.}
}
\description{
Plots a line for each entry of a \code{tf}-object.
\code{geom_spaghetti} does spaghetti plots, \code{geom_meatballs} does spaghetti plots
with points for the actual evaluations.
}
\section{\code{y} aesthetic}{

Mandatory. Used to designate a column of class \code{tf} to be visualized.
}

\examples{
set.seed(1221)
data = data.frame(col = sample(gl(5, 2)))
data$f = rgp(10)
data$fi = jiggle(data$f)
data$fb = tfb(data$f)
library(ggplot2)
ggplot(data, aes(y = f, color = depth(f))) + geom_spaghetti()
ggplot(data, aes(y = fi, shape = col, color = col)) + geom_meatballs()
ggplot(data, aes(y = fi)) + geom_meatballs(spaghetti = FALSE) + 
  facet_wrap(~col)
}
\keyword{datasets}
