% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{jiggle}
\alias{jiggle}
\title{Make a \code{tfd} function (more) irregular}
\usage{
jiggle(f, ...)
}
\arguments{
\item{f}{a \code{tfd} object}

\item{..}{not used currently}
}
\description{
... by moving around its \code{arg}-values.
}
