% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-dataframe.R
\name{tf_evaluate.data.frame}
\alias{tf_evaluate.data.frame}
\title{Evaluate \code{tf}s inside a \code{data.frame}}
\usage{
\method{tf_evaluate}{data.frame}(object, ..., arg)
}
\arguments{
\item{object}{a \code{data.frame}-like object with \code{tf} columns.}

\item{...}{optional: a selection of \code{tf}-columns. If empty, all \code{tf}-variables
in the data frame are selected. You can supply bare variable names,
select all variables between \code{x} and \code{z} with \code{x:z}, exclude \code{y} with \code{-y}.
For more options, see the \code{\link[dplyr:select]{dplyr::select()}} documentation.}

\item{arg}{optional evaluation grid (vector or list of vectors).
Defaults to \code{tf_arg(object)}.}
}
\value{
Replaces \code{tf}-columns with list columns of
smaller \code{data.frames} containing the functions' arguments (\code{arg}) and
evaluations (\code{value}) and returns the modified nested dataframe.
}
\description{
Evaluate \code{tf}s inside a \code{data.frame}
}
\details{
The \code{arg}-argument of \code{tf_evaluate.data.frame} method can be a
list of \code{arg}-vectors or -lists used as the \code{arg} argument for the
\code{\link[tf:tf_evaluate]{tf::tf_evaluate()}}-method for the respective \code{tf}-columns in \code{object}.
\code{...} is not used for a \code{tf}-\code{object}, but a second unnamed argument to these
methods will be interpreted as \code{arg}.
}
