% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb-fpc.R
\name{tfb_fpc}
\alias{tfb_fpc}
\alias{tfb_fpc.data.frame}
\alias{tfb_fpc.matrix}
\alias{tfb_fpc.numeric}
\alias{tfb_fpc.tf}
\title{Constructors/converters for functional data in FPC-basis representation}
\usage{
tfb_fpc(data, ...)

\method{tfb_fpc}{data.frame}(data, id = 1, arg = 2, value = 3,
  domain = NULL, smooth = TRUE, resolution = NULL, ...)

\method{tfb_fpc}{matrix}(data, arg = NULL, domain = NULL, smooth = TRUE,
  resolution = NULL, ...)

\method{tfb_fpc}{numeric}(data, arg = NULL, domain = NULL, smooth = TRUE,
  resolution = NULL, ...)

\method{tfb_fpc}{tf}(data, arg = NULL, smooth = TRUE, ...)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} or \code{list} of suitable shape, or another \code{tf}-object.}

\item{...}{arguments to the call to \code{\link[refund:fpca.sc]{refund::fpca.sc()}} that computes the
(regularized/smoothed) FPCA. Unless set by the user \code{tidyfun} uses \code{pve = .995} to
determine the truncation levels and uses \code{bs = 15} basis functions for
the mean function and the marginal bases for the covariance surface.}

\item{id}{The name/number of the column defining which data belong to which function.}

\item{arg}{\code{numeric}, or list of \code{numeric}s. The evaluation grid. See Details.
For the \code{data.frame}-method: the name/number of the column defining the evaluation grid.}

\item{value}{The name/number of the column containing the function evaluations.}

\item{domain}{range of the \code{arg}.}

\item{smooth}{use smoothed mean function and smoothed covariance surface estimates
from \code{\link[refund:fpca.sc]{refund::fpca.sc()}} or simply perform a (truncated) PCA of the data? See Details.}

\item{resolution}{resolution of the evaluation grid. See details for \code{\link[=tfd]{tfd()}}.}
}
\description{
These functions perform a (functional) principal component analysis of the
input data and return an \code{tfb} \code{tf}-object that uses the empirical
eigenfunctions as basis functions for representing the data. By default, a
\code{smooth}ed FPCA via \code{\link[refund:fpca.sc]{refund::fpca.sc()}} is used to compute eigenfunctions and
scores based on the smoothed empirical covariance.
If \code{smooth =FALSE}, a fast, unregularized PCA of the data is done instead.
}
\details{
ATM, the unsmoothed version does not use orthogonal basis "functions"
for irregular / non-equidistant grids and does not work for
incomplete/irregular data.
}
\seealso{
tfb
}
