% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brackets.R
\name{[.tf}
\alias{[.tf}
\alias{[<-.tf}
\title{Accessing/evaluating, subsetting and subassigning \code{tfs}}
\usage{
\method{[}{tf}(x, i, j, interpolate = TRUE, matrix = TRUE)

\method{[}{tf}(x, i) <- value
}
\arguments{
\item{x}{an \code{tf}}

\item{i}{index of the observations (\code{integer}ish, \code{character} or \code{logical},
usual R rules apply)}

\item{j}{The \code{argvals} used to evaluate the functions. A (list of) \code{numeric}
vectors.}

\item{interpolate}{should functions be evaluated (i.e., inter-/extrapolated)
for \code{argvals} for which no original data is available? Only relevant for
\code{tfd}, defaults to TRUE.}

\item{matrix}{should the result be returned as a \code{matrix} or as a list of
\code{data.frame}s? If TRUE, \code{j} has to be a (list of a) single vector of
\code{argvals}. See return value.}

\item{value}{\code{tf} object for subassignment. This is (currently) very strictly typed,
i.e. only objects that are of the same class and have compatible \code{argvals} can be
subassigned.}
}
\value{
If \code{j} is missing, a subset of the functions in \code{x} as given by
\code{i}.\cr If \code{j} is given and \code{matrix == TRUE}, a numeric matrix of function
evaluations in which each row represents one function and each column
represents one \code{argval} as given in argument \code{j}, with an attribute
\code{argvals}=\code{j} and row- and column-names derived from \code{x[i]} and \code{j}.\cr If
\code{j} is given and \code{matrix == FALSE}, a list of \code{tbl_df}s with columns
\code{argvals} = \code{j} and \code{data} = evaluations at \code{j} for each observation in
\code{i}.
}
\description{
These functions access, subset, replace and evaluate \code{tfs}.
For more information on creating \code{tf}s and converting them to/from
\code{list}, \code{data.frame} or \code{matrix}, see \code{\link[=tfd]{tfd()}} and \code{\link[=fbase]{fbase()}}. \cr
Note that these break certain R conventions for vector-like objects:\cr
\itemize{
\item no argument recycling,
\item no indexing with \code{NA}, names not present in \code{x} or integers \code{> length(x)}
}
}
\details{
Both of these will yield errors. Subassigning new elements to positions
beyond the original length of the object will still fill up the vector with \code{NAs},
though.
}
