% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{deriv.tfd}
\alias{deriv.tfd}
\alias{deriv.tfb}
\alias{integrate}
\alias{integrate.function}
\alias{integrate.tfd}
\alias{integrate.tfb}
\title{Derivatives and integrals of functional data}
\usage{
\method{deriv}{tfd}(expr, order = 1, arg = NULL, ...)

\method{deriv}{tfb}(expr, order = 1, ...)

integrate(f, lower, upper, ...)

\method{integrate}{function}(f, lower, upper, ..., subdivisions = 100L,
  rel.tol = .Machine$double.eps^0.25, abs.tol = rel.tol,
  stop.on.error = TRUE, keep.xy = FALSE, aux = NULL)

\method{integrate}{tfd}(f, lower = domain(f)[1], upper = domain(f)[2],
  definite = TRUE, arg, ...)

\method{integrate}{tfb}(f, lower = domain(f)[1], upper = domain(f)[2],
  definite = TRUE, arg, ...)
}
\arguments{
\item{expr}{a \code{tf}}

\item{order}{order of differentiation. Maximally 2 for \code{tfb} with \code{mgcv}-spline bases.}

\item{arg}{grid to use for the differentiation/integration. Not the \code{arg} of the returned object.}

\item{...}{not used}

\item{f}{an \R function taking a numeric first argument and returning
    a numeric vector of the same length.  Returning a non-finite element will
    generate an error.}

\item{lower}{lower limits of the integration range. For \code{definite=TRUE}, this can be
a vector of the same length as \code{f}.}

\item{upper}{upper limits of the integration range (but see \code{definite} arg / Description).
For \code{definite=TRUE}, this can be a vector of the same length as \code{f}.}

\item{subdivisions}{the maximum number of subintervals.}

\item{rel.tol}{relative accuracy requested.}

\item{abs.tol}{absolute accuracy requested.}

\item{stop.on.error}{logical. If true (the default) an error stops the
    function.  If false some errors will give a result with a warning in
    the \code{message} component.}

\item{keep.xy}{unused.  For compatibility with S.}

\item{aux}{unused.  For compatibility with S.}

\item{definite}{should the definite integral  be returned (default)
or the antiderivative. See Description.}
}
\value{
a \code{tf} with (slightly) different \code{arg} (and \code{basis})
}
\description{
Derivatives of \code{tf}s use finite differences of the evaluations
for \code{tfd} and finite differences of the basis functions for \code{tfb}.
Note that, for some spline bases like \code{"cr"} or \code{"tp"} which always begin/end linearly,
computing second derivatives will produce artefacts at the outer limits
of the functions' domain due to these boundary constraints. Basis \code{"bs"} does
not have this problem, but tends to yield slightly less stable fits.

Integration of \code{tf}s is done by quadrature (trapezoid rule, specifically).
By default the scalar definite integral \eqn{\int^{upper}_{lower}f(s)ds} is returned
(option \code{definite = TRUE}), alternatively for \code{definite = FALSE} something
like the \emph{anti-derivative} on \code{[lower, upper]}, e.g. an \code{tfd} object
representing \eqn{F(t) = \int^{t}_{lower}f(s)ds}, for \eqn{t \in}\code{[lower, upper]},
is returned.
}
\details{
\code{integrate.function} is simply a wrapper for \code{\link[stats:integrate]{stats::integrate()}}.
}
