% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-format.R
\name{print.tf}
\alias{print.tf}
\alias{print.feval_reg}
\alias{print.feval_irreg}
\alias{print.fbase}
\alias{format.tf}
\title{Pretty printing and formatting for functional data}
\usage{
\method{print}{tf}(x, n = 10, ...)

\method{print}{feval_reg}(x, n = 10, ...)

\method{print}{feval_irreg}(x, n = 10, ...)

\method{print}{fbase}(x, n = 10, ...)

\method{format}{tf}(x, digits = 2, nsmall = 0, ...)
}
\arguments{
\item{x}{any \R object (conceptually); typically numeric.}

\item{n}{how many elements of \code{x} to print out}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{how many significant digits are to be used for
    numeric and complex \code{x}.  The default, \code{NULL}, uses
    \code{\link{getOption}("digits")}.  This is a suggestion: enough decimal
    places will be used so that the smallest (in magnitude) number has
    this many significant digits, and also to satisfy \code{nsmall}.
    (For the interpretation for complex numbers see \code{\link{signif}}.)}

\item{nsmall}{the minimum number of digits to the right of the decimal
    point in formatting real/complex numbers in non-scientific formats.
    Allowed values are \code{0 <= nsmall <= 20}.}
}
\description{
Print/format \code{tf}-objects.
}
