% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R, R/math.R
\name{Ops.tf}
\alias{Ops.tf}
\alias{==.tfd}
\alias{!=.tfd}
\alias{==.tfb}
\alias{!=.tfb}
\alias{Ops.tfd}
\alias{Ops.tfb}
\alias{Math.tfd}
\alias{Math.tfb}
\alias{cummax.tfd}
\alias{cummin.tfd}
\alias{cumsum.tfd}
\alias{cumprod.tfd}
\alias{cummax.tfb}
\alias{cummin.tfb}
\alias{cumsum.tfb}
\alias{cumprod.tfb}
\alias{Summary.tf}
\title{Math, Summary and Ops Methods for \code{tf}}
\usage{
\method{Ops}{tf}(e1, e2)

\method{==}{tfd}(e1, e2)

\method{!=}{tfd}(e1, e2)

\method{==}{tfb}(e1, e2)

\method{!=}{tfb}(e1, e2)

\method{Ops}{tfd}(e1, e2)

\method{Ops}{tfb}(e1, e2)

\method{Math}{tfd}(x, ...)

\method{Math}{tfb}(x, ...)

\method{cummax}{tfd}(...)

\method{cummin}{tfd}(...)

\method{cumsum}{tfd}(...)

\method{cumprod}{tfd}(...)

\method{cummax}{tfb}(...)

\method{cummin}{tfb}(...)

\method{cumsum}{tfb}(...)

\method{cumprod}{tfb}(...)

\method{Summary}{tf}(...)
}
\arguments{
\item{e1}{an \code{tf} or a numeric vector}

\item{e2}{an \code{tf} or a numeric vector}

\item{x}{an \code{tf}}

\item{...}{\code{tf}-objects (not used for \code{Math} group generic)}
}
\description{
These define methods and operators that mostly work \code{argval}-wise on
\code{tf} objects, see \code{?groupGeneric} for implementation details.
}
\details{
See examples below. Equality checks of functional objects are rather iffy and
not very reliable at this point. Note that \code{max} and \code{min} are not guaranteed
to be maximal/minmal over the entire domain, only on the evaluation grid used for
computation. With the exception of addition and multiplication,
operations on \code{tfb}-objects first evaluate them over their \code{arg},
perform computations on these evaluations and then convert back to an \code{tfb}-
object, so a loss of precision should be expected, especially so if bases are small
or data is very wiggly.
}
\examples{
set.seed(1859)
f = tf_rgp(4)
2 * f == f + f
sum(f) == f[1] + f[2] + f[3] + f[4]
log(exp(f)) == f
plot(f, points = FALSE)
lines(range(f), col = 2, lty = 2)

f2 = tfb(tf_rgp(5), k = 50)
layout(t(1:2))
plot(f2, col = 1:5)
plot(cumsum(f2), col = 1:5)
# use ?tf_integrate.tfd for "function-wise" integrals i.e., weighted cumulative sums...
lines(f2)
}
