% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tf_arg}
\alias{tf_arg}
\alias{tf_evaluations}
\alias{tf_count}
\alias{tf_domain}
\alias{tf_evaluator}
\alias{tf_basis}
\alias{tf_evaluator<-}
\alias{arg<-}
\alias{arg<-.tfd_irreg}
\alias{arg<-.tfd_reg}
\alias{tf_resolution}
\alias{coef.tfb}
\alias{rev.tf}
\alias{is.na.tf}
\alias{is.na.tfd_irreg}
\alias{is_tf}
\alias{is_irreg}
\alias{is_tfd}
\alias{is_tfb}
\title{Utility functions for \code{tf}-objects}
\usage{
tf_arg(f)

tf_evaluations(f)

tf_count(f)

tf_domain(f)

tf_evaluator(f, forget = FALSE)

tf_basis(f, as_tfd = FALSE, forget = FALSE)

tf_evaluator(x) <- value

arg(x) <- value

\method{arg}{tfd_irreg}(x) <- value

\method{arg}{tfd_reg}(x) <- value

tf_resolution(f)

\method{coef}{tfb}(object, ...)

\method{rev}{tf}(x)

\method{is.na}{tf}(x)

\method{is.na}{tfd_irreg}(x)

is_tf(x)

is_irreg(x)

is_tfd(x)

is_tfb(x)
}
\arguments{
\item{f}{an \code{tf} object}

\item{forget}{extract the evaluator or basis-creating function without its cache?
See \code{\link[memoise:forget]{memoise::forget()}}. Defaults to \code{FALSE}.}

\item{as_tfd}{should the basis be returned as a \code{tfd} evaluated on \code{tf_arg(f)}? Defaults to FALSE.}

\item{x}{an \code{tf} object}

\item{value}{\strong{for \code{evaluator=}:} (bare or quoted) name of a function that
can be used to interpolate an \code{tfd}. Needs to accept vector arguments \code{x},
\code{arg}, \code{evaluations} and return evaluations of the function defined by
\code{arg}, \code{evaluations} at \code{x}
\strong{for \code{arg=}:} a (list of) new \code{arg}-values}

\item{object}{as usual}

\item{...}{dots}
}
\description{
A bunch of methods & utilities that do what they say: extract or set the
respective attributes of a \code{tf}-object.
}
