% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/utils.R
\name{arg}
\alias{arg}
\alias{evaluations}
\alias{n_evaluations}
\alias{domain}
\alias{evaluator}
\alias{basis}
\alias{evaluator<-}
\alias{arg<-}
\alias{arg<-.tfd_irreg}
\alias{arg<-.tfd_reg}
\alias{coef.tfb}
\alias{rev.tf}
\alias{is_tf}
\alias{is_irreg}
\alias{is_tfd}
\alias{is_tfb}
\alias{is.na.tf}
\title{Utility functions for \code{tf}-objects}
\usage{
arg(f)

evaluations(f)

n_evaluations(f)

domain(f)

evaluator(f, forget = FALSE)

basis(f)

evaluator(x) <- value

arg(x) <- value

\method{arg}{tfd_irreg}(x) <- value

\method{arg}{tfd_reg}(x) <- value

\method{coef}{tfb}(object, ...)

\method{rev}{tf}(x)

is_tf(x)

is_irreg(x)

is_tfd(x)

is_tfb(x)

\method{is.na}{tf}(x)
}
\arguments{
\item{f}{an \code{tf} object}

\item{forget}{extract the evaluator function without its cache?
See \code{\link[memoise:::forget]{memoise::::forget()}}Defaults to \code{FALSE}.}

\item{x}{an \code{tf} object}

\item{value}{\strong{for \code{evaluator<-}:} name of a function that can be used to interpolate an \code{tfd}. Needs to
accept vector arguments \code{x}, \code{arg}, \code{evaluations} and return
evaluations of the function defined by \code{arg}, \code{evaluations} at \code{x}
\strong{for \code{arg<-}:} a (list of) new \code{arg}-values}

\item{object}{as usual}

\item{...}{dots}
}
\description{
A bunch of methods & utilities that do what they say.
}
