% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/utils.R
\name{argvals}
\alias{argvals}
\alias{evaluations}
\alias{n_evaluations}
\alias{domain}
\alias{evaluator}
\alias{basis}
\alias{evaluator<-}
\alias{argvals<-}
\alias{coef.tfb}
\alias{rev.tf}
\alias{is_tf}
\alias{is_irreg}
\alias{is_tfd}
\alias{is_tfb}
\alias{is.na.tf}
\title{Utility functions for \code{tf}-objects}
\usage{
argvals(f)

evaluations(f)

n_evaluations(f)

domain(f)

evaluator(f)

basis(f)

evaluator(x) <- value

argvals(x) <- value

\method{coef}{tfb}(object, ...)

\method{rev}{tf}(x)

is_tf(x)

is_irreg(x)

is_tfd(x)

is_tfb(x)

\method{is.na}{tf}(x)
}
\arguments{
\item{f}{an \code{tf} object}

\item{x}{an \code{tf} object}

\item{value}{for \code{evaluator<-}: name of a function that can be used to interpolate an \code{tfd}. Needs to
accept vector arguments \code{x}, \code{argvals}, \code{evaluations} and return
evaluations of the function defined by \code{argvals}, \code{evaluations} at \code{x}
for \code{argvals<-}: a list of grid points, for internal use only.}

\item{object}{as usual}

\item{...}{dots}
}
\description{
A bunch of methods & utilities that do what they say.
}
