% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-ggplot.R
\name{transform_tf_data}
\alias{transform_tf_data}
\title{Transform data with tf aesthetics into long format}
\usage{
transform_tf_data(data, tf_aesthetics, arg = NULL, interpolate = TRUE)
}
\arguments{
\item{data}{Data frame containing tf objects}

\item{tf_aesthetics}{List of tf aesthetic mappings}

\item{arg}{Optional evaluation grid}

\item{interpolate}{Whether to interpolate tf objects}
}
\value{
Transformed data frame in long format
}
\description{
Transform data with tf aesthetics into long format
}
