% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-vctrs.R
\name{vec_cast.tfd_reg}
\alias{vec_cast.tfd_reg}
\alias{vec_cast.tfd_irreg}
\alias{vec_cast.tfd_reg.tfd_reg}
\alias{vec_cast.tfd_reg.tfd_irreg}
\alias{vec_cast.tfd_irreg.tfd_reg}
\alias{vec_cast.tfd_irreg.tfd_irreg}
\alias{vctrs}
\alias{vec_ptype2.tfd_reg}
\alias{vec_ptype2.tfd_reg.tfd_reg}
\alias{vec_ptype2.tfd_reg.tfd_irreg}
\alias{vec_ptype2.tfd_irreg}
\alias{vec_ptype2.tfd_irreg.tfd_reg}
\alias{vec_ptype2.tfd_irreg.tfd_irreg}
\alias{vec_ptype2.tfd.tfd}
\alias{vec_ptype2.tfb}
\alias{vec_ptype2.tfb.tfb}
\title{vctrs methods for \code{tf} objects}
\usage{
\method{vec_cast}{tfd_reg}(x, to, ...)

\method{vec_cast}{tfd_irreg}(x, to, ...)

\method{vec_cast.tfd_reg}{tfd_reg}(x, to, ...)

\method{vec_cast.tfd_reg}{tfd_irreg}(x, to, ...)

\method{vec_cast.tfd_irreg}{tfd_reg}(x, to, ...)

\method{vec_cast.tfd_irreg}{tfd_irreg}(x, to, ...)

\method{vec_ptype2}{tfd_reg}(x, y, ...)

\method{vec_ptype2.tfd_reg}{tfd_reg}(x, y, ...)

\method{vec_ptype2.tfd_reg}{tfd_irreg}(x, y, ...)

\method{vec_ptype2}{tfd_irreg}(x, y, ...)

\method{vec_ptype2.tfd_irreg}{tfd_reg}(x, y, ...)

\method{vec_ptype2.tfd_irreg}{tfd_irreg}(x, y, ...)

\method{vec_ptype2}{tfd.tfd}(x, y, ...)

\method{vec_ptype2}{tfd_reg}(x, y, ...)

\method{vec_ptype2.tfb}{tfb}(x, y, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}

\item{y}{Vector types.}
}
\description{
These functions are the extensions that allow \code{tidyfun} vectors
to work with \code{vctrs}.
}
