% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-vctrs.R
\name{vctrs}
\alias{vctrs}
\alias{vec_ptype2.tfd}
\alias{vec_ptype2.tfb}
\alias{vec_ptype2.tfd.tfd}
\alias{vec_ptype2.tfb.tfb}
\alias{vec_cast.tfd}
\alias{vec_cast.tfb}
\alias{vec_cast.tfd.tfd}
\alias{vec_cast.tfb.tfb}
\alias{vec_cast.tfb.tfd}
\alias{vec_cast.tfd.tfb}
\title{vctrs methods for tf objects}
\usage{
vec_ptype2.tfd(x, y, ...)

vec_ptype2.tfb(x, y, ...)

\method{vec_ptype2.tfd}{tfd}(x, y, ...)

\method{vec_ptype2.tfb}{tfb}(x, y, ...)

vec_cast.tfd(x, to, ...)

vec_cast.tfb(x, to, ...)

\method{vec_cast}{tfd.tfd}(x, to, ...)

\method{vec_cast}{tfb.tfb}(x, to, ...)

\method{vec_cast}{tfb.tfd}(x, to, ...)

\method{vec_cast}{tfd.tfb}(x, to, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{x_arg, y_arg}{Argument names for \code{x} and \code{y}.}
}
\description{
vctrs methods for tf objects
}
